requirements.txt
openai
langchain
chromadb
python-dotenv
PyMuPDF
tqdm
pytesseract
pdf2image
pdfminer.six
Pillow
unstructured

streamlit
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
openai
PyPDF2
